fix(hw-42): separate svg, separate navigarion element
