+++ my-app/src/Components/ThemeButton/ThemeButton.tsx
@@ -1,6 +1,8 @@
 import { useContext } from "react"
 import { ThemeContext } from "../Context/themeContext"
 import './ThemeButton.css'
+import { ReactComponent as Sun } from './sun.svg';
+import { ReactComponent as Moon } from './moon.svg';
 
 type Props = { changeThemeClick: () => void }
 
@@ -8,24 +10,6 @@ export const ThemeButton = ({ changeThemeClick }: Props) => {
     const theme = useContext(ThemeContext)
 
     return <button className='changeThemeButton' onClick={changeThemeClick}>
-        {theme === 'light' ?
-            <svg width="50px" height="50px" viewBox="0 0 50 50">
-                <path d="M 20.46875 2.539062 C 10.773438 4.578125 3.554688 12.539062 2.429688 22.453125 C 2.265625 23.9375 2.363281 27.195312 2.617188 28.613281 C 3.054688 31.015625 3.601562 32.703125 4.6875 34.914062 C 5.683594 36.960938 7.402344 39.421875 8.71875 40.703125 C 9.140625 41.113281 9.148438 41.125 9.21875 40.898438 C 9.34375 40.429688 9.394531 39.09375 9.289062 38.476562 C 9.199219 37.960938 9.210938 37.8125 9.328125 37.625 C 9.5625 37.285156 9.902344 37.335938 10.226562 37.765625 C 10.371094 37.96875 10.8125 38.445312 11.179688 38.828125 C 12.304688 39.960938 13.601562 40.460938 15.382812 40.449219 C 15.917969 40.449219 16.5625 40.398438 16.816406 40.351562 C 17.453125 40.226562 17.714844 40.371094 17.753906 40.851562 C 17.78125 41.179688 17.753906 41.230469 17.324219 41.523438 C 16.425781 42.148438 15.75 42.765625 15.351562 43.339844 C 14.929688 43.9375 14.550781 44.726562 14.550781 44.992188 C 14.550781 45.1875 16.335938 46.015625 17.910156 46.53125 C 22.703125 48.144531 27.792969 48.105469 32.609375 46.445312 C 37.460938 44.765625 41.765625 41.308594 44.523438 36.875 C 45.527344 35.265625 46.503906 33.015625 47.0625 31.054688 C 47.363281 29.96875 47.703125 28.328125 47.625 28.261719 C 47.609375 28.234375 47.390625 28.46875 47.167969 28.78125 C 46.523438 29.640625 44.6875 31.425781 43.703125 32.15625 C 40.578125 34.472656 36.828125 35.734375 32.9375 35.773438 L 31.542969 35.789062 L 31.289062 36.28125 C 30.929688 36.960938 30.773438 37.109375 30.460938 37.109375 C 30.109375 37.109375 29.851562 36.757812 29.890625 36.308594 C 29.941406 35.59375 29.960938 35.605469 29.003906 35.421875 C 26.945312 35.03125 24.304688 33.9375 22.558594 32.753906 C 18.09375 29.757812 15.203125 25.226562 14.40625 19.96875 C 14.210938 18.703125 14.210938 15.890625 14.40625 14.628906 C 15.097656 10.046875 17.453125 5.867188 20.898438 3.09375 C 21.296875 2.773438 21.671875 2.46875 21.726562 2.421875 C 21.867188 2.296875 21.425781 2.34375 20.46875 2.539062 Z M 20.46875 2.539062 " />
-                <path d="M 29.015625 18.789062 C 28.953125 18.859375 28.835938 19.140625 28.757812 19.40625 C 28.476562 20.351562 27.910156 20.898438 26.894531 21.230469 C 26.53125 21.347656 26.367188 21.453125 26.328125 21.609375 C 26.257812 21.875 26.503906 22.265625 26.738281 22.265625 C 26.835938 22.265625 27.140625 22.382812 27.421875 22.519531 C 28.039062 22.820312 28.46875 23.378906 28.679688 24.160156 C 28.886719 24.875 29.171875 25.109375 29.539062 24.863281 C 29.679688 24.765625 29.824219 24.539062 29.890625 24.296875 C 30.1875 23.171875 30.8125 22.570312 31.828125 22.402344 C 32.285156 22.335938 32.546875 22.078125 32.5 21.765625 C 32.46875 21.523438 32.207031 21.367188 31.5625 21.191406 C 31.335938 21.132812 30.984375 20.945312 30.789062 20.773438 C 30.429688 20.449219 29.980469 19.640625 29.980469 19.289062 C 29.980469 18.78125 29.304688 18.429688 29.015625 18.789062 Z M 29.015625 18.789062 " />
-                <path d="M 31.757812 30.175781 C 31.132812 30.78125 30.359375 31.230469 29.589844 31.453125 C 29.34375 31.523438 29.140625 31.582031 29.121094 31.59375 C 29.109375 31.601562 29.277344 31.804688 29.492188 32.039062 C 30 32.597656 30.507812 33.574219 30.710938 34.382812 L 30.867188 35.019531 L 31.476562 34.433594 C 32.117188 33.796875 32.84375 33.390625 33.640625 33.183594 C 33.914062 33.117188 34.148438 33.054688 34.171875 33.039062 C 34.1875 33.027344 33.976562 32.765625 33.691406 32.453125 C 33.117188 31.828125 32.695312 31.007812 32.519531 30.148438 C 32.460938 29.84375 32.402344 29.589844 32.390625 29.589844 C 32.382812 29.601562 32.101562 29.863281 31.757812 30.175781 Z M 31.757812 30.175781 " />
-                <path d="M 10.351562 39.726562 C 10.351562 41.035156 9.578125 42.898438 8.585938 43.984375 C 8.273438 44.328125 8.007812 44.628906 8.007812 44.648438 C 8.007812 44.65625 8.367188 44.6875 8.820312 44.695312 C 9.257812 44.71875 9.84375 44.773438 10.109375 44.835938 C 10.890625 45.007812 12.140625 45.65625 12.726562 46.191406 C 13.015625 46.453125 13.28125 46.679688 13.300781 46.679688 C 13.328125 46.679688 13.378906 46.320312 13.40625 45.867188 C 13.507812 44.484375 14.101562 43.171875 15.164062 42.050781 C 15.46875 41.726562 15.722656 41.453125 15.722656 41.445312 C 15.722656 41.4375 15.273438 41.40625 14.726562 41.375 C 13.203125 41.296875 11.875 40.773438 10.820312 39.84375 L 10.351562 39.433594 Z M 10.351562 39.726562 " />
-            </svg> :
-            <svg width="50px" height="50px" viewBox="0 0 50 50" fill="#fff">
-                <path d="M 24.5 2.078125 C 24.316406 2.195312 24.09375 2.460938 23.992188 2.65625 C 23.820312 3.007812 23.820312 3.066406 23.976562 4.070312 C 24.082031 4.726562 24.160156 6.007812 24.1875 7.421875 C 24.238281 9.921875 24.265625 10.019531 24.824219 10.3125 C 25.4375 10.625 26.269531 10.320312 26.53125 9.6875 C 26.828125 8.992188 26.5625 3.945312 26.171875 2.519531 C 25.984375 1.867188 25.148438 1.648438 24.5 2.078125 Z M 24.5 2.078125 " />
-                <path d="M 8.820312 7.03125 C 8.203125 7.402344 8.015625 8.203125 8.429688 8.757812 C 8.546875 8.925781 9.042969 9.539062 9.523438 10.117188 C 10.007812 10.703125 10.859375 11.804688 11.414062 12.570312 C 12.226562 13.671875 12.5 13.984375 12.792969 14.101562 C 13.351562 14.328125 14.070312 14.109375 14.394531 13.632812 C 14.734375 13.09375 14.628906 12.832031 13.40625 11.171875 C 12.179688 9.484375 10.8125 7.78125 10.234375 7.207031 C 9.804688 6.765625 9.335938 6.71875 8.820312 7.03125 Z M 8.820312 7.03125 " />
-                <path d="M 39.421875 9.226562 C 39.265625 9.304688 38.59375 9.84375 37.910156 10.429688 C 37.226562 11.015625 36.171875 11.894531 35.566406 12.390625 C 34.960938 12.890625 34.421875 13.378906 34.375 13.476562 C 34.132812 13.9375 34.34375 14.773438 34.785156 15.078125 C 35.304688 15.4375 35.976562 15.234375 36.960938 14.40625 C 41.296875 10.78125 41.375 10.703125 41.242188 10.136719 C 41.140625 9.765625 40.851562 9.414062 40.488281 9.238281 C 40.078125 9.042969 39.773438 9.042969 39.421875 9.226562 Z M 39.421875 9.226562 " />
-                <path d="M 23.632812 14.171875 C 21.984375 14.289062 20.683594 14.640625 19.199219 15.390625 C 15.734375 17.128906 13.242188 20.46875 12.597656 24.21875 C 12.441406 25.125 12.441406 27.195312 12.585938 27.976562 C 13.242188 31.28125 15.226562 33.859375 18.242188 35.34375 C 19.980469 36.203125 21.796875 36.570312 23.828125 36.484375 C 25.878906 36.398438 27.703125 35.90625 29.6875 34.921875 C 33.203125 33.164062 35.507812 30.578125 36.484375 27.296875 C 36.710938 26.542969 36.75 26.21875 36.757812 25.046875 C 36.765625 23.429688 36.582031 22.585938 35.878906 21.09375 C 34.296875 17.742188 31.59375 15.382812 28.507812 14.648438 C 26.933594 14.289062 24.929688 14.082031 23.632812 14.171875 Z M 23.378906 16.835938 C 23.71875 16.992188 23.847656 16.992188 24.601562 16.882812 C 28.144531 16.386719 31.90625 18.417969 33.585938 21.75 C 34.304688 23.183594 34.539062 25.1875 34.148438 26.660156 C 33.496094 29.171875 31.40625 31.398438 28.398438 32.804688 C 25.292969 34.257812 22.585938 34.484375 20.019531 33.484375 C 18.976562 33.085938 18.132812 32.507812 17.1875 31.542969 C 15.578125 29.902344 14.804688 28.085938 14.804688 25.929688 C 14.804688 21.855469 18.183594 17.773438 22.414062 16.75 C 22.675781 16.679688 22.921875 16.632812 22.949219 16.640625 C 22.976562 16.648438 23.171875 16.738281 23.378906 16.835938 Z M 23.378906 16.835938 " />
-                <path d="M 0.617188 22.949219 C 0.320312 23.066406 0 23.601562 0 23.976562 C 0 24.382812 0.304688 24.914062 0.65625 25.117188 C 0.859375 25.242188 1.921875 25.410156 4.511719 25.734375 C 6.484375 25.976562 8.234375 26.171875 8.398438 26.171875 C 8.625 26.171875 8.789062 26.085938 8.992188 25.867188 C 9.523438 25.304688 9.296875 24.210938 8.574219 23.875 C 8.390625 23.789062 6.6875 23.535156 4.65625 23.289062 C 0.722656 22.8125 0.917969 22.820312 0.617188 22.949219 Z M 0.617188 22.949219 " />
-                <path d="M 42.1875 23.429688 C 39.902344 23.484375 39.757812 23.523438 39.382812 24.082031 C 39.160156 24.433594 39.179688 25.136719 39.421875 25.46875 C 39.746094 25.90625 40.058594 25.945312 42.773438 25.867188 C 44.140625 25.828125 46.132812 25.800781 47.195312 25.789062 C 49.375 25.78125 49.570312 25.734375 49.851562 25.136719 C 50.136719 24.53125 49.882812 23.757812 49.304688 23.46875 C 49.0625 23.339844 46.386719 23.320312 42.1875 23.429688 Z M 42.1875 23.429688 " />
-                <path d="M 35.085938 34.277344 C 34.71875 34.5 34.472656 34.929688 34.472656 35.359375 C 34.472656 35.625 34.589844 35.789062 35.1875 36.367188 C 35.578125 36.75 37.0625 38.210938 38.484375 39.617188 C 41.085938 42.195312 41.152344 42.246094 41.75 42.1875 C 42.265625 42.140625 42.773438 41.492188 42.773438 40.890625 C 42.773438 40.554688 42.675781 40.410156 41.757812 39.421875 C 40.085938 37.609375 36.777344 34.460938 36.257812 34.1875 C 35.9375 34.023438 35.4375 34.0625 35.085938 34.277344 Z M 35.085938 34.277344 " />
-                <path d="M 10.039062 38.40625 C 8.898438 39.226562 7.179688 40.4375 6.21875 41.09375 C 5.046875 41.914062 4.445312 42.390625 4.335938 42.597656 C 4.03125 43.171875 4.304688 44.070312 4.863281 44.328125 C 5.398438 44.578125 5.644531 44.472656 7.570312 43.132812 C 10.292969 41.25 13.105469 39.210938 13.320312 38.984375 C 13.742188 38.507812 13.59375 37.570312 13.027344 37.117188 C 12.5 36.699219 12.296875 36.789062 10.039062 38.40625 Z M 10.039062 38.40625 " />
-                <path d="M 23.171875 38.671875 C 23.007812 38.769531 22.8125 38.953125 22.753906 39.0625 C 22.625 39.296875 22.609375 43.484375 22.714844 46.046875 L 22.78125 47.507812 L 23.117188 47.820312 C 23.390625 48.085938 23.523438 48.144531 23.886719 48.144531 C 24.132812 48.144531 24.460938 48.078125 24.617188 47.988281 C 25.175781 47.703125 25.1875 47.597656 25.125 43.40625 C 25.085938 41.035156 25.03125 39.472656 24.960938 39.257812 C 24.734375 38.574219 23.828125 38.273438 23.171875 38.671875 Z M 23.171875 38.671875 " />
-            </svg>
-        }
+        {theme === 'light' ? <Moon /> : <Sun />}
     </button>
 }
