--- my-app/src/Components/ThemeButton/images.js
@@ -1,74 +1,0 @@
-export const sun = <svg version="1.0" xmlns="http://www.w3.org/2000/svg"
-width="512.000000pt" height="512.000000pt" viewBox="0 0 512.000000 512.000000"
-preserveAspectRatio="xMidYMid meet">
-
-<g transform="translate(0.000000,512.000000) scale(0.100000,-0.100000)"
-fill="#000000" stroke="none">
-<path d="M2509 4907 c-19 -12 -42 -39 -52 -59 -18 -36 -18 -42 -2 -145 11 -67
-19 -198 22 -343 5 -256 8 -266 65 -296 63 -32 148 -1 175 64 30 71 3 588 -37
-734 -19 67 -105 89 -171 45z"/>
-<path d="M903 4400 c-63 -38 -82 -120 -40 -177 12 -17 63 -80 112 -139 50 -60
-137 -173 194 -251 83 -113 111 -145 141 -157 57 -23 131 -1 164 48 35 55 24
-82 -101 252 -126 173 -266 347 -325 406 -44 45 -92 50 -145 18z"/>
-<path d="M4037 4175 c-16 -8 -85 -63 -155 -123 -70 -60 -178 -150 -240 -201
--62 -51 -117 -101 -122 -111 -25 -47 -3 -133 42 -164 53 -37 122 -16 223 69
-444 371 452 379 438 437 -10 38 -40 74 -77 92 -42 20 -73 20 -109 1z"/>
-<path d="M2420 3669 c-169 -12 -302 -48 -454 -125 -355 -178 -610 -520 -676
--904 -16 -93 -16 -305 -1 -385 67 -338 270 -602 579 -754 178 -88 364 -126
-572 -117 210 9 397 59 600 160 360 180 596 445 696 781 23 77 27 110 28 230 1
-166 -18 252 -90 405 -162 343 -439 585 -755 660 -161 37 -366 58 -499 49z
-m-26 -273 c35 -16 48 -16 125 -5 363 51 748 -157 920 -498 74 -147 98 -352 58
--503 -67 -257 -281 -485 -589 -629 -318 -149 -595 -172 -858 -70 -107 41 -193
-100 -290 199 -165 168 -244 354 -244 575 0 417 346 835 779 940 27 7 52 12 55
-11 3 -1 23 -10 44 -20z"/>
-<path d="M63 2770 c-30 -12 -63 -67 -63 -105 0 -42 31 -96 67 -117 21 -13 130
--30 395 -63 202 -25 381 -45 398 -45 23 0 40 9 61 31 54 58 31 170 -43 204
--19 9 -193 35 -401 60 -403 49 -383 48 -414 35z"/>
-<path d="M4320 2721 c-234 -6 -249 -10 -287 -67 -23 -36 -21 -108 4 -142 33
--45 65 -49 343 -41 140 4 344 7 453 8 223 1 243 6 272 67 29 62 3 141 -56 171
--25 13 -299 15 -729 4z"/>
-<path d="M3593 1610 c-38 -23 -63 -67 -63 -111 0 -27 12 -44 73 -103 40 -39
-192 -189 338 -333 266 -264 273 -269 334 -263 53 5 105 71 105 133 0 34 -10
-49 -104 150 -171 186 -510 508 -563 536 -33 17 -84 13 -120 -9z"/>
-<path d="M1028 1187 c-117 -84 -293 -208 -391 -275 -120 -84 -182 -133 -193
--154 -31 -59 -3 -151 54 -177 55 -26 80 -15 277 122 279 193 567 402 589 425
-43 49 28 145 -30 191 -54 43 -75 34 -306 -132z"/>
-<path d="M2373 1160 c-17 -10 -37 -29 -43 -40 -13 -24 -15 -453 -4 -715 l7
--150 34 -32 c28 -27 42 -33 79 -33 25 0 59 7 75 16 57 29 58 40 52 469 -4 243
--10 403 -17 425 -23 70 -116 101 -183 60z"/>
-</g>
-</svg>
-
-export const moon = <svg version="1.0" xmlns="http://www.w3.org/2000/svg"
-width="1280.000000pt" height="1074.000000pt" viewBox="0 0 1280.000000 1074.000000"
-preserveAspectRatio="xMidYMid meet">
-<g transform="translate(0.000000,1074.000000) scale(0.100000,-0.100000)"
-fill="#000000" stroke="none">
-<path d="M7295 10684 c203 -193 404 -398 512 -524 681 -788 1121 -1762 1262
--2795 108 -781 39 -1576 -200 -2335 -568 -1799 -2034 -3166 -3871 -3610 -292
--70 -634 -124 -880 -137 -65 -3 -121 -10 -124 -14 -2 -4 28 -34 68 -66 810
--657 1821 -1070 2858 -1167 445 -42 498 -43 845 -16 753 56 1382 226 2030 545
-593 292 1069 642 1536 1131 603 630 1061 1459 1288 2334 203 783 227 1534 75
-2350 -234 1254 -947 2421 -1964 3216 -931 727 -2043 1114 -3280 1140 l-215 5
-60 -57z"/>
-<path d="M4428 8979 c-20 -35 -59 -108 -88 -161 -28 -54 -55 -100 -58 -102 -4
--2 -83 -12 -177 -21 -93 -10 -173 -21 -177 -24 -4 -4 49 -67 118 -140 101
--106 125 -137 120 -154 -22 -93 -67 -329 -63 -334 3 -3 76 29 163 71 86 42
-162 76 168 76 6 0 81 -41 166 -90 85 -49 156 -88 157 -86 2 1 -8 83 -22 182
-l-24 179 32 33 c18 19 78 74 135 125 56 50 102 93 102 97 0 4 -80 21 -177 39
--98 18 -181 35 -184 39 -4 4 -35 70 -69 147 -94 212 -81 199 -122 124z"/>
-<path d="M605 8031 c-89 -123 -136 -180 -149 -181 -10 0 -109 29 -219 65 -110
-36 -201 64 -203 62 -2 -2 55 -83 126 -181 72 -98 130 -183 130 -188 0 -8 -177
--260 -243 -346 -5 -7 -7 -15 -4 -18 2 -3 97 25 211 61 113 37 209 64 214 61 5
--3 67 -86 138 -183 l129 -178 5 229 5 229 215 72 c194 66 212 74 188 83 -14 6
--107 36 -205 68 -98 31 -184 61 -190 66 -9 7 -13 72 -15 235 l-3 224 -130
--180z"/>
-<path d="M4412 6004 c-123 -153 -230 -282 -237 -286 -8 -5 -155 46 -354 121
--188 71 -341 126 -341 122 0 -4 88 -141 195 -305 107 -163 195 -304 195 -312
-0 -8 -101 -141 -225 -296 -124 -154 -224 -283 -222 -285 3 -2 161 39 352 91
-191 52 352 93 358 90 7 -2 102 -142 212 -310 l200 -306 6 69 c4 37 12 202 19
-366 10 211 17 301 26 308 7 6 166 52 354 104 188 51 344 96 346 98 5 5 -58 30
--431 171 -127 48 -236 91 -244 95 -11 7 -10 64 4 358 9 192 16 358 14 368 -3
-13 -69 -62 -227 -261z"/>
-</g>
-</svg>
